rpi/app/requirements.txt
annotated-types==0.5.0
blis==0.7.11
boltons==23.0.0
brotlipy==0.7.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
conda==23.5.2
conda-content-trust==0.1.3
conda-libmamba-solver==23.5.0
conda-package-handling==2.1.0
conda_package_streaming==0.8.0
confection==0.1.3
contourpy==1.1.1
cryptography==39.0.1
cycler==0.11.0
cymem==2.0.8
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.4
flickrapi==2.4.0
fonttools==4.42.1
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.4.5
langcodes==3.3.0
libmambapy==1.4.1
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==10.0.1
pip==23.1.2
pluggy==1.0.0
preshed==3.0.9
pycosat==0.6.4
pycparser==2.21
pydantic==2.4.1
pydantic_core==2.10.1
pyOpenSSL==23.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
ruamel.yaml==0.17.21
scikit-learn==1.3.1
scipy==1.10.1
setuptools==67.8.0
six==1.16.0
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.1.12
threadpoolctl==3.2.0
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wasabi==1.1.2
wheel==0.38.4
zipp==3.17.0
zstandard==0.19.0
